Current Constants
I went through and took 256 ADC readings at 1/4 amp intervals from zero amps to 4 amps. I then obtained an average ADC value for each of the currents and ploted them. I then took and calculated a line between zero and 4 amps and ploted this. both apeared to be aproximately the same so I took the values found and updated the constants used to calculate the current.